Move default component to its own file
